refactor(PersonName): 과한 원시값 포장 제거
