+++ src/main/java/model/people/Person.java
@@ -1,13 +1,23 @@
 package model.people;
 
 import java.util.Objects;
-import model.Name;
 
 public class Person {
-    private final Name name;
+    private static final int MIN_NAME_LENGTH = 1;
+    private static final int MAX_NAME_LENGTH = 5;
+
+    private final String name;
 
     public Person(final String rawName) {
-        this.name = new PersonName(rawName);
+        validateNameLength(rawName);
+        this.name = rawName;
+    }
+
+    private void validateNameLength(final String rawName) {
+        final int length = rawName.length();
+        if (length < MIN_NAME_LENGTH || length > MAX_NAME_LENGTH) {
+            throw new IllegalArgumentException("이름은 최소 1글자 최대 5글자여야 합니다.");
+        }
     }
 
     @Override
@@ -28,6 +38,6 @@ public class Person {
     }
 
     public String getName() {
-        return name.getName();
+        return name;
     }
 }
