+++ src/test/java/model/people/PersonCountTest.java
@@ -23,4 +23,12 @@ class PersonCountTest {
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("참여 인원은 최소 2여야 합니다.");
     }
+
+    @ParameterizedTest(name = "사람의 이름은 최소 1글자 최대 5글자다.")
+    @ValueSource(strings = {"", "mollly"})
+    void createPersonThrowException(String name) {
+        assertThatThrownBy(() -> new Person(name))
+                .isExactlyInstanceOf(IllegalArgumentException.class)
+                .hasMessage("이름은 최소 1글자 최대 5글자여야 합니다.");
+    }
 }
